Distinguish different "no data" errors returned by API
Define two additional errors (InvalidBusinessParameterError and
InvalidPSRTypeError). They are fundamentally different from the
NoMatchingDataError. NoMatchingDataError is returned whenever a
particular period has no data associated to it (specific hours/days).
The other two errors are returned whenever the chosen PSR makes no
sense, or if the zone identifier is wrong, which is an error independent
of the chosen period.