Fixes (#50)
* Avoid truncation error

In rare cases, generation per unit is not retrieved in the right order,
and an error is raised. Sorting along the index prior to truncation
prevents this error.

* Raise NoMatchingDataError

If *all* the chunked requests returned a NoMatchingDataError, we should
still raise a NoMatchingDataError instead of silently returning an emtpy
dataframe.

* Add DK-ENERGINET Control Area Code

Actual Generation per Unit requires a Control Area identifier to be
used: add this identifier for DK.

* Distinguish different "no data" errors returned by API

Define two additional errors (InvalidBusinessParameterError and
InvalidPSRTypeError). They are fundamentally different from the
NoMatchingDataError. NoMatchingDataError is returned whenever a
particular period has no data associated to it (specific hours/days).
The other two errors are returned whenever the chosen PSR makes no
sense, or if the zone identifier is wrong, which is an error independent
of the chosen period.