Fix intervals in yearly_limited
Yearly_limited decorator splits queries into year long parts and
truncates resulting subqueries. The intervals for truncation are
right-closed except for the (chronologically) first one - which is
closed. The logic that handles this special case is faulty and breaks if
the queried start timestamp isn't found in the data.

This commit fixes the faulty assumption by checking for the first
interval more explicitly.